+++ reeltime-front/src/App.js
@@ -3,18 +3,17 @@ import './App.css';
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Login from './components/Login';
 import Signup from './components/Signup';
-import Searchbox from './components/Searchbox';
+import Home from './components/Home';
 
 function App() {
   return (
     <div>
       <BrowserRouter>
-        {/* <Login></Login> */}
         <Routes>
           <Route path="/signup" element={<Signup />}/>
           <Route exact path="/login" element={<Login />}/>
+          <Route path="/home" element={<Home />}/>
         </Routes>
-        <Searchbox></Searchbox>
       </BrowserRouter>
     </div>
   );
