+++ reeltime-front/src/components/Home.css
@@ -1,3 +1,7 @@
+.home-title {
+    text-align: center;
+}
+
 .search-box-container {
     position: absolute;
     top: 20px;
