+++ reeltime-front/src/components/Searchbox.js
@@ -1,23 +1,23 @@
-import React, { useState } from "react";
-import "./Searchbox.css";
+// import React, { useState } from "react";
+// import "./Searchbox.css";
 
-function Searchbox({ value, setSearchValue }) {
-  const handleChange = (event) => {
-    setSearchValue(event.target.value);
-  };
+// function Searchbox({ value, setSearchValue }) {
+//   const handleChange = (event) => {
+//     setSearchValue(event.target.value);
+//   };
 
-  return (
-    <div className="search-box-container">
-      <i className="search-box-icon fas fa-search"></i>
-      <input
-        className="search-box-input"
-        value={value}
-        onChange={handleChange}
-        placeholder="Search Movies..."
-      />
-    </div>
-  );
-}
+//   return (
+//     <div className="search-box-container">
+//       <i className="search-box-icon fas fa-search"></i>
+//       <input
+//         className="search-box-input"
+//         value={value}
+//         onChange={handleChange}
+//         placeholder="Search Movies..."
+//       />
+//     </div>
+//   );
+// }
 
 
-export default Searchbox
+// export default Searchbox
